cdk_python_test/pipeline_stack.py
from aws_cdk import Stack
from aws_cdk import (
    aws_codecommit as codecommit,
)
from constructs import Construct


class WorkshopPipelineStack(Stack):
    def __init__(self, scope: Construct, id: str, **kwargs) -> None:
        super().__init__(scope, id, **kwargs)

        # Creates a CodeCommit repository called 'WorkshopRepo'
        repo = codecommit.Repository(
            self, "WorkshopRepo", repository_name="WorkshopRepo"
        )

        # Pipeline code goes here
